backend/requirements.txt
# FastAPI core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Pydantic v2
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# AI (Claude)
anthropic==0.26.0

# Redis
redis==5.0.4
aioredis==2.0.1

# HTTP client
httpx==0.27.0
aiohttp==3.9.5

# Google OAuth
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0

# Rate limiting
slowapi==0.1.9

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.1
tenacity==8.3.0
